restructure apps
