--- src/pretix/api/__init__.py
@@ -1,12 +1,0 @@
-from django.apps import AppConfig
-
-
-class PretixApiConfig(AppConfig):
-    name = 'pretix.api'
-    label = 'pretixapi'
-
-    def ready(self):
-        from . import signals, webhooks  # noqa
-
-
-default_app_config = 'pretix.api.PretixApiConfig'
