+++ src/pretix/helpers/apps.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig
+
+class PretixHelpersConfig(AppConfig):
+    name = 'pretix.helpers'
+    label = 'pretixhelpers'
+
+
+default_app_config = 'pretix.helpers.PretixHelpersConfig'
